Update build system by running autoreconf -fvi.
This patch is the result of running

   ./autoreconf -fvi

with autoreconf v2.71.  It updates many of the auto-generated files.
The only hand-written files that have been changed is INSTALL, which
now gives detailed info about how to use configure, instead of just
forwarding to README.

More changes:
* The generated COPYING file has been discarded since there is LICENSE already.
* The generated compile script has been added.
* A .gitignore file has been added.  It ignores the auto-generated
  autom4te.cache/ folder.

Running bootstrap with this patch comes up with on changes.

This patch has been tested to configure and build on:
* x86_64 GNU/Linux  (AMD64, LP64), i.e. int is 32-bit, GCC 13.2.
* ppc64le GNU/Linux (PowerPC little endian), GCC 11.5.
* ppc64 GNU/Linux   (PowerPC big endian), GCC 15.2.  There is the known
  problem that TestU01 has no proper prototypes and GCC 15 uses C23,
  hence the build was run as  make CFLAGS='std=c99'.